MixedModelsExtras.jl API

Coefficient of Determination

StatsBase.r2 — Function

r2(model::StatisticalModel)
r²(model::StatisticalModel)

Coefficient of determination (R-squared).

For a linear model, the R² is defined as $ESS/TSS$, with $ESS$ the explained sum of squares and $TSS$ the total sum of squares.

r2(model::StatisticalModel, variant::Symbol)
r²(model::StatisticalModel, variant::Symbol)

Pseudo-coefficient of determination (pseudo R-squared).

For nonlinear models, one of several pseudo R² definitions must be chosen via variant. Supported variants are:

:MacFadden (a.k.a. likelihood ratio index), defined as $1 - \log (L)/\log (L_0)$;

:CoxSnell, defined as $1 - (L_0/L)^{2/n}$;

:Nagelkerke, defined as $(1 - (L_0/L)^{2/n})/(1 - L_0^{2/n})$.

:devianceratio, defined as $1 - D/D_0$.

In the above formulas, $L$ is the likelihood of the model, $L_0$ is the likelihood of the null model (the model with only an intercept), $D$ is the deviance of the model (from the saturated model), $D_0$ is the deviance of the null model, $n$ is the number of observations (given by nobs).

The Cox-Snell and the deviance ratio variants both match the classical definition of R² for linear models.

r2(model::LinearMixedModel; conditional=false)
r²(model::LinearMixedModel; conditional=false)

Coefficient of determination (R-squared).

R² is very non trivial for mixed models for a number of reasons and the measures here are particularly simplistic. For conditional=true, the Pearson correlation between the fitted and observed values is simply squared, in line with the relationship between the coefficient of determination and the Pearson correlation for classical OLS models. For conditional=false, new predicted values are generated based on only the fixed-effects estimates and the correlation between these predictions and the observed values is again squared.

In both cases, it is important to note that despite the usual naming convention (R² for the coefficient of determination and r for Pearson's correlation coefficient), these quantitaties are not defined in terms of the other. Instead, the correlation is a standardized measure of covariance and the coefficient of determination is measured relative to a null model. (The total sum of squares is in some sense the squared residual error of the null model.) Even generalizations of the coefficient of determination that define this value in terms of the likelihoods of an intercept-only model (i.e. the sample mean) and the likelihood of the fitted model fail to generalize to the mixed-model case: should "intercept-only" mean truly just the fixed effect (i.e. comparing the likelihoods of a classical OLS and a mixed-effects model)? or should it also include an random intercept for each grouping variable? What is the null model in this case, when determining whether each grouping variable contributes to overall fit? There is no single, clear, agreed-upon answer for these questions. The bottom line is that there are many potential ways to define a coefficient of determination for linear mixed models (and even more for the generalzed case which combines all the problems of R² for GLM and for LMM) and none of them have all the properties of the coefficient of determination for classical OLS regression.

There are more advanced approximations (see e.g. Nakagawa and colleagues' 2013 2017 papers) but the fundamental philosophical issues above remain. In the future, additional approximations may be supported, but this is not a high priority. Pull requests are welcome.

For more information, see the GLMM FAQ

StatsBase.adjr2 — Function

adjr2(model::StatisticalModel)
adjr²(model::StatisticalModel)

Adjusted coefficient of determination (adjusted R-squared).

For linear models, the adjusted R² is defined as $1 - (1 - (1-R^2)(n-1)/(n-p))$, with $R^2$ the coefficient of determination, $n$ the number of observations, and $p$ the number of coefficients (including the intercept). This definition is generally known as the Wherry Formula I.

adjr2(model::StatisticalModel, variant::Symbol)
adjr²(model::StatisticalModel, variant::Symbol)

Adjusted pseudo-coefficient of determination (adjusted pseudo R-squared).

For nonlinear models, one of the several pseudo R² definitions must be chosen via variant. The only currently supported variants are :MacFadden, defined as $1 - (\log (L) - k)/\log (L0)$ and :devianceratio, defined as $1 - (D/(n-k))/(D_0/(n-1))$. In these formulas, $L$ is the likelihood of the model, $L0$ that of the null model (the model including only the intercept), $D$ is the deviance of the model, $D_0$ is the deviance of the null model, $n$ is the number of observations (given by nobs) and $k$ is the number of consumed degrees of freedom of the model (as returned by dof).

adjr2(model::LinearMixedModel; conditional=false)
adjr²(model::LinearMixedModel; conditional=false)

Adjusted coefficient of determination (adjusted R-squared).

For linear models, the adjusted R² is defined as $(1 - (1-R^2)(n-1)/(n-p))$, with $R^2$ the coefficient of determination, $n$ the number of observations, and $p$ the number of coefficients (including the intercept). This definition is generally known as the Wherry Formula I.

For mixed-models, we use this same adjustment for the R² value computed with r2. conditional whether or not the random effects are taken into account when computing the quality of the model fit. They are nonetheless included in the adjustment (i.e. as number $p$ of parameters).

R² is not a clearly defined concept for mixed models, nor does it have all the properties typically expected of the coefficient of determination. See r2 for more information.

Intra-Class Correlation Coefficient

MixedModelsExtras.icc — Function

icc(model::MixedModel, [groups])

Compute the intra-class correlation coefficient (ICC) for a mixed model.

The ICC is defined as the variance attributable to the groups divided by the total variance from all groups and the observation-level (residual) variance. In other words, the ICC can be interpreted as the proportion of the variance explainable by the grouping/nesting structure.

A single group can be specified as a Symbol, e.g. :subj or a number of groups can be specified as an array: [:subj, :item]. If no groups are specified, then all grouping variables are used.

Note

The value returned here is sometimes called the "adjusted ICC" and does not take the variance of the fixed effects into account (the "conditional ICC").

Note

The result returned aggregates across groups. If you require the ICC for each group separately, then you must call icc separately for each group.