MixedModelsExtras.jl API

Coefficient of Determination

StatsAPI.r2 — Function

r2(model::StatisticalModel)
r²(model::StatisticalModel)

Coefficient of determination (R-squared).

For a linear model, the R² is defined as $ESS/TSS$, with $ESS$ the explained sum of squares and $TSS$ the total sum of squares.

StatsAPI.adjr2 — Function

adjr2(model::StatisticalModel)
adjr²(model::StatisticalModel)

Adjusted coefficient of determination (adjusted R-squared).

For linear models, the adjusted R² is defined as $1 - (1 - (1-R^2)(n-1)/(n-p))$, with $R^2$ the coefficient of determination, $n$ the number of observations, and $p$ the number of coefficients (including the intercept). This definition is generally known as the Wherry Formula I.

Intra-Class Correlation Coefficient

MixedModelsExtras.icc — Function

icc(model::MixedModel, [groups])

Compute the intra-class correlation coefficient (ICC) for a mixed model.

The ICC is defined as the variance attributable to the groups divided by the total variance from all groups and the observation-level (residual) variance. In other words, the ICC can be interpreted as the proportion of the variance explainable by the grouping/nesting structure.

A single group can be specified as a Symbol, e.g. :subj or a number of groups can be specified as an array: [:subj, :item]. If no groups are specified, then all grouping variables are used.

Note

The value returned here is sometimes called the "adjusted ICC" and does not take the variance of the fixed effects into account (the "conditional ICC").

Note

The result returned aggregates across groups. If you require the ICC for each group separately, then you must call icc separately for each group.

Variance Inflation Factor

MixedModelsExtras.vif — Function

vif(m::RegressionModel)

Compute the variance inflation factor (VIF).

Returns a vector of inflation factors computed for each coefficient except for the intercept. In other words, the corresponding coefficients are coefnames(m)[2:end].

The variance inflation factor (VIF) measures the increase in the variance of a parameter's estimate in a model with multiple parameters relative to the variance of a parameter's estimate in a model containing only that parameter.

See also coefnames, gvif.

Warning

This method will fail if there is (numerically) perfect multicollinearity, i.e. rank deficiency (in the fixed effects). In that case though, the VIF isn't particularly informative anyway.

MixedModelsExtras.termnames — Function

termnames(model)

Return the names associated with (fixed effects) terms in a model.

For models with only continuous predictors, this is the same as coefnames(model). For models with categorical predictors, the returned names reflect the categorical predictor and not the coefficients resulting from the choice of contrast coding.

MixedModelsExtras.gvif — Function

gvif(m::RegressionModel; scale=false)

Compute the generalized variance inflation factor (GVIF).

If scale=true, then each GVIF is scaled by the degrees of freedom for (number of coefficients associated with) the predictor: $GVIF^(1 / (2*df))$

Returns a vector of inflation factors computed for each term except for the intercept. In other words, the corresponding coefficients are termnames(m)[2:end].

The generalized variance inflation factor (VIF) measures the increase in the variance of a (group of) parameter's estimate in a model with multiple parameters relative to the variance of a parameter's estimate in a model containing only that parameter. For continuous, numerical predictors, the GVIF is the same as the VIF, but for categorical predictors, the GVIF provides a single number for the entire group of contrast-coded coefficients associated with a categorical predictor.

See also termnames, vif.

Warning

This method will fail if there is (numerically) perfect multicollinearity, i.e. rank deficiency (in the fixed effects). In that case though, the VIF isn't particularly informative anyway.

References

Fox, J., & Monette, G. (1992). Generalized Collinearity Diagnostics. Journal of the American Statistical Association, 87(417), 178. doi:10.2307/2290467